+++ src/controllers/index.ts
@@ -2,7 +2,7 @@ export * from './tag.controller';
 export * from './json-model.controller';
 export * from './pclass.controller';
 export * from './pfunction.controller';
-export * from './pclass-instance.controller';
+export * from './pclassi.controller';
 export * from './openapi.controller';
 export * from './project.controller';
 export * from './package.controller';
