+++ src/models/index.ts
@@ -2,7 +2,7 @@ export * from './json-model.model';
 export * from './tag.model';
 export * from './pfunction.model';
 export * from './pclass.model';
-export * from './pclass-instance.model';
+export * from './pclassi.model';
 export * from './openapi.model';
 export * from './storage.model';
 export * from './project.model';
