+++ src/repositories/index.ts
@@ -2,7 +2,7 @@ export * from './json-model.repository';
 export * from './tag.repository';
 export * from './pfunction.repository';
 export * from './pclass.repository';
-export * from './pclass-instance.repository';
+export * from './pclassi.repository';
 export * from './openapi.repository';
 export * from './project.repository';
 export * from './package.repository';
